Improved support for the table of contents.
Removed a bug:
Some automatically generated anchor names were not unique.

Clicking on a url will now highlight that url in the table of contents.

Removed a bug:
The hash was being ignored when you loaded a new page.
(The hash already worked when you clicked on the table on contents.)