Hide the second scroll bar as needed.
If the screen is wide enough,
the user can choose to have the table of contents
in it's own container,
with it's own scrollbar.
In this case it's always visible.

If the screen is too narrow,
that option goes away and only one scrollbar is visible.
The table of contents and the body will share a scrollbar.